app/requirements.txt
google-adk>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
firecrawl-py>=0.2.0
ipykernel>=6.25.0
langchain-community>=0.2.0
litellm>=1.40.0
markitdown[all]>=0.0.1
openai>=1.30.0
google-generativeai>=0.7.0
typing-extensions>=4.5.0 